Complete Refactor/Update to Cox
