+++ runelite-client/src/main/java/net/runelite/client/plugins/coxhelper/OlmCrippleTimerOverlay.java
@@ -23,7 +23,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.zcox;
+package net.runelite.client.plugins.coxhelper;
 
 
 import java.awt.Color;
